% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.summary.R
\name{my.summary}
\alias{my.summary}
\title{my.summary}
\usage{
my.summary(x)
}
\arguments{
\item{x}{output from my.lm function}
}
\value{
no return value - this function only cats a summary of the regression
}
\description{
cats regression summary from output of my.lm function
}
\examples{
get(data(mtcars))
my.fit = my.lm(mpg ~ cyl, data = mtcars)
summary(my.fit)

}
