% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.lm.R
\name{my.lm}
\alias{my.lm}
\title{my.lm}
\usage{
my.lm(formula, data, weights)
}
\arguments{
\item{formula}{formula specifying the regression equation}

\item{data}{data.frame to be used to fit the model}

\item{weights}{optional numeric vector to be used in a WLS regression}
}
\value{
A list containing the coefficients, residuals, fitted values, and more
}
\description{
Fits a linear regression model using OLS
}
\examples{
get(data(mtcars))
my.fit = my.lm(mpg ~ cyl, data = mtcars)
my.fit$coefficients

}
